python/pyproject.toml
[project]
name = "codereview-agent"
version = "0.1.0"
description = "AI-powered CodeReview Agent with LangChain + LangGraph"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.6.0",
    "mypy>=1.0.0",
    "matplotlib>=3.0.0",
]

[project.scripts]
codereview = "codereview.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python-version = "3.9"
strict = true
